Refactoring for map room types
Created puppet scenes
Added WIP map icons
Created BattleConfig
