+++ scenes/Puppets/Enemies/Boss1.tscn
@@ -1,0 +1,22 @@
+[gd_scene load_steps=4 format=3 uid="uid://bi5iqbwpsd381"]
+
+[ext_resource type="Script" path="res://scenes/Puppets/scripts/PuppetTemplate.cs" id="1_310x1"]
+[ext_resource type="Texture2D" uid="uid://veedngaorx3l" path="res://scenes/BattleDirector/assets/Enemy1.png" id="2_bxcx7"]
+[ext_resource type="PackedScene" uid="uid://bgomxovxs7sr8" path="res://scenes/Puppets/HealthBar.tscn" id="3_otelj"]
+
+[node name="EnemPuppet" type="Node2D" node_paths=PackedStringArray("_healthBar", "Sprite")]
+script = ExtResource("1_310x1")
+_healthBar = NodePath("ProgressBar")
+Sprite = NodePath("Sprite")
+StartPos = Vector2(500, 125)
+InitScale = Vector2(2, 2)
+
+[node name="Sprite" type="Sprite2D" parent="."]
+position = Vector2(0, -32)
+texture = ExtResource("2_bxcx7")
+
+[node name="ProgressBar" parent="." instance=ExtResource("3_otelj")]
+offset_left = -75.0
+offset_top = 32.0
+offset_right = 75.0
+offset_bottom = 52.0
